# fix: stm tests run time regression