--- a/docs/devbook/bump-versions/bump_versions.sh
+++ b/docs/devbook/bump-versions/bump_versions.sh
@@ -211,7 +211,7 @@ else
   if [ true = $COMMIT ]
   then
     git add --update $OPEN_API_FILE Cargo.lock ./*/Cargo.toml ./internal/*/Cargo.toml ./mithril-test-lab/*/Cargo.toml
-    git add --update ./*/package.json ./*/package-lock.json examples/*/package.json examples/*/package-lock.json
+    git add --update ./*/package.json ./*/package-lock.json mithril-client-wasm/ci-test/package-lock.json examples/*/package.json examples/*/package-lock.json
     git commit -m "$COMMIT_MESSAGE"
   fi
 fi