--- a/mithril-client-wasm/Cargo.toml
+++ b/mithril-client-wasm/Cargo.toml
@@ -17,7 +17,7 @@ anyhow = "1.0.95"
 async-trait = "0.1.86"
 chrono = { version = "0.4.39", features = ["serde"] }
 futures = "0.3.31"
-mithril-client = { path = "../mithril-client", default-features = false, features = ["unstable", "num-integer-backend"] }
+mithril-client = { path = "../mithril-client", default-features = false, features = ["unstable"] }
 serde = { version = "1.0.217", features = ["derive"] }
 serde-wasm-bindgen = "0.6.5"
 serde_json = "1.0.138"