--- a/mithril-client-wasm/Cargo.toml
+++ b/mithril-client-wasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mithril-client-wasm"
-version = "0.8.1"
+version = "0.8.2"
 description = "Mithril client WASM"
 authors = { workspace = true }
 edition = { workspace = true }