--- a/mithril-client-wasm/package.json
+++ b/mithril-client-wasm/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mithril-dev/mithril-client-wasm",
-  "version": "0.8.1",
+  "version": "0.8.2",
   "description": "Mithril client WASM",
   "license": "Apache-2.0",
   "collaborators": [